+++ bmds/drunner.py
@@ -1,10 +1,12 @@
 import os
 from simple_settings import settings
+import subprocess
 
 from . import utils, session
+from .models.base import RunStatus
 
 
-class BatchDfileRunner(object):
+class BatchDfileRunner:
     """
     Batch-execute a list of pre-created d-files.
 
@@ -15,8 +17,6 @@ class BatchDfileRunner(object):
     def __init__(self, inputs):
         self.tempfiles = utils.TempFileList()
         self.inputs = inputs
-        self.outputs = []
-        self.execute()
 
     def get_outfile(self, dfile, model_name):
         outfile = dfile.replace('.(d)', '.out')
@@ -38,42 +38,64 @@ class BatchDfileRunner(object):
         outfile = os.path.join(path, prefix + fn)
         return outfile
 
+    def execute_job(self, obj):
+        """
+        Execute the BMDS model and parse outputs if successful.
+        """
+
+        # get executable path
+        exe = session.BMDS\
+            .get_model(obj['bmds_version'], obj['model_name'])\
+            .get_exe_path()
+
+        # write dfile
+        dfile = self.tempfiles.get_tempfile(prefix='bmds-dfile-', suffix='.(d)')
+        with open(dfile, 'w') as f:
+            f.write(obj['dfile'])
+
+        outfile = self.get_outfile(dfile, obj['model_name'])
+        oo2 = outfile.replace('.out', '.002')
+
+        proc = subprocess.Popen(
+            [exe, dfile],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE)
+
+        output = None
+        stdout = ''
+        stderr = ''
+
+        try:
+            stdout, stderr = proc.communicate(
+                timeout=settings.BMDS_MODEL_TIMEOUT_SECONDS)
+
+            if os.path.exists(outfile):
+                with open(outfile, 'r') as f:
+                    output = f.read()
+
+            status = RunStatus.SUCCESS.value
+            stdout = stdout.decode().strip()
+            stderr = stderr.decode().strip()
+
+        except subprocess.TimeoutExpired:
+            proc.kill()
+            status = RunStatus.FAILURE.value
+            stdout, stderr = proc.communicate()
+
+        finally:
+            if os.path.exists(outfile):
+                self.tempfiles.append(outfile)
+            if os.path.exists(oo2):
+                self.tempfiles.append(oo2)
+
+        self.tempfiles.cleanup()
+
+        return dict(
+            status=status,
+            output=output,
+            stdout=stdout,
+            stderr=stderr,
+        )
+
     def execute(self):
-
-        for obj in self.inputs:
-
-            # get executable path
-            exe = session.BMDS.get_model(obj['bmds_version'],
-                                         obj['model_name']).get_exe_path()
-
-            # write dfile
-            dfile = self.tempfiles.get_tempfile(prefix='bmds-dfile-', suffix='.(d)')
-            with open(dfile, 'w') as f:
-                f.write(obj['dfile'])
-
-            output = {
-                'output_created': False,
-                'outfile': None,
-            }
-            try:
-                utils.RunBMDS(
-                    [exe, dfile],
-                    timeout=settings.BMDS_MODEL_TIMEOUT_SECONDS
-                ).call()
-                outfile = self.get_outfile(dfile, obj['model_name'])
-                oo2 = outfile.replace('.out', '.002')
-                if os.path.exists(outfile):
-                    self.tempfiles.append(outfile)
-                    output['output_created'] = True
-                    with open(outfile, 'r') as f:
-                        output['outfile'] = f.read()
-                if os.path.exists(oo2):
-                    self.tempfiles.append(oo2)
-            except Exception as e:
-                raise e
-            finally:
-                self.tempfiles.cleanup()
-
-            self.outputs.append(output)
-
-        return self.outputs
+        return list(map(self.execute_job, self.inputs))