+++ bmds/logic/recommender.py
@@ -1,11 +1,12 @@
 from .. import constants
 from . import rules
 
+import pandas as pd
+from simple_settings import settings
+
 
 class Recommender(object):
     # Recommendation logic for a specified data-type.
-
-    SUFFICIENTLY_CLOSE_BMDL = 3
 
     def __init__(self, dtype, overrides=None):
 
@@ -21,6 +22,7 @@ class Recommender(object):
 
         self.dtype = dtype
         self.rules = self._get_rule_defaults(**rule_args)
+        self.SUFFICIENTLY_CLOSE_BMDL = settings.SUFFICIENTLY_CLOSE_BMDL
 
         if overrides:
             raise NotImplementedError('Overrides not implemented (yet).')
@@ -64,7 +66,7 @@ class Recommender(object):
             rules.BmdBmdlRatio(
                 failure_bin=constants.BIN_NO_CHANGE,
                 threshold=5.,
-                rule_name='BMD/BMDL (warning)',
+                rule_name='BMD to BMDL ratio (warning)',
             ),
             rules.BmdBmdlRatio(
                 failure_bin=constants.BIN_WARNING,
@@ -170,6 +172,13 @@ class Recommender(object):
 
     def show_rules(self):
         return u'\n'.join([rule.__unicode__() for rule in self.rules])
+
+    def rules_df(self):
+        df = pd.DataFrame(
+            data=[rule.as_row() for rule in self.rules],
+            columns=['rule_name', 'enabled', 'failure_bin', 'threshold'])
+        df = df[df.enabled == True]
+        return df
 
     def _get_bmdl_ratio(self, models):
         """Return BMDL ratio in list of models."""