+++ bmds/monkeypatch.py
@@ -23,7 +23,7 @@ import sys
 from simple_settings import settings
 
 from .session import BMDS
-from .models.base import BMDModel
+from .models.base import BMDModel, RunStatus
 from .exceptions import RemoteBMDSExcecutionException
 
 
@@ -80,10 +80,15 @@ if platform.system() != 'Windows':
 
         return _request_session
 
-    def _set_outputs(model, result):
-        model.output_created = result['output_created']
-        if model.output_created:
-            model.parse_results(result['outfile'])
+    def _set_results(model, results=None):
+        if results is None:
+            model._set_job_outputs(RunStatus.DID_NOT_RUN)
+        else:
+            status = results.pop('status')
+            if status == RunStatus.SUCCESS:
+                model._set_job_outputs(RunStatus.SUCCESS, **results)
+            elif status == RunStatus.FAILURE:
+                model._set_job_outputs(RunStatus.FAILURE)
 
     def execute_model(self):
         # execute single model
@@ -94,12 +99,10 @@ if platform.system() != 'Windows':
             payload = _get_payload([self])
             logger.debug('Submitting payload: {}'.format(payload))
             resp = session.post(url, data=payload)
-            result = resp.json()[0]
+            results = resp.json()[0]
         else:
-            result = {'output_created': False}
-
-        self.execution_end = datetime.now()
-        _set_outputs(self, result)
+            results = None
+        _set_results(self, results)
 
     def execute_session(self):
         # submit data
@@ -110,7 +113,7 @@ if platform.system() != 'Windows':
             if model.can_be_executed:
                 executable_models.append(model)
             else:
-                _set_outputs(model, {'output_created': False})
+                _set_results(model)
 
         if len(executable_models) == 0:
             return
@@ -122,11 +125,9 @@ if platform.system() != 'Windows':
         resp = session.post(url, data=payload)
 
         # parse results for each model
-        end_time = datetime.now()
         jsoned = resp.json()
-        for model, result in zip(executable_models, jsoned):
-            model.execution_end = end_time
-            _set_outputs(model, result)
+        for model, results in zip(executable_models, jsoned):
+            _set_results(model, results)
 
     # print startup error if host is None
     if settings.BMDS_HOST is None or \