+++ bmds/settings.py
@@ -5,6 +5,17 @@ import platform
 # before closing (in seconds)
 BMDS_MODEL_TIMEOUT_SECONDS = 10
 
+# Maximum polynomial used when setting default models; generally only used when
+# the number of dose groups is large in a dataset. This is so that if you have
+# 15 dose-groups you won't try to model using a 14 degree polynomial model.
+# maximum_poly =  min(dataset dose groups - 1, MAXIMUM_POLYNOMIAL_ORDER)
+MAXIMUM_POLYNOMIAL_ORDER = 8
+
+# In model recommendation; if range of BMDLs is less than 3x different, the
+# model with the smallest AIC will be recommended, otherwise the model with the
+# smallest BMDL will be recommended.
+SUFFICIENTLY_CLOSE_BMDL = 3
+
 # Only required if running BMDS on a non-Windows computer
 if platform.system() != 'Windows':
 
@@ -23,6 +34,9 @@ LOGGING = {
     'formatters': {
         'default': {
             'format': '%(asctime)s %(levelname)s %(name)s %(message)s'
+        },        
+        'simple': {
+            'format': '%(levelname)s %(message)s'
         },
     },
     'handlers': {
@@ -33,6 +47,11 @@ LOGGING = {
             'maxBytes': 50 * 1024 * 1024,
             'backupCount': 10,
             'formatter': 'default'
+        },
+        'console': {
+            'level': 'DEBUG',
+            'class': 'logging.StreamHandler',
+            'formatter': 'simple',
         },
     },
     'loggers': {