+++ bmds/utils.py
@@ -1,10 +1,8 @@
 import ctypes
 import logging
-from subprocess import Popen
 import os
 import sys
 import tempfile
-import threading
 
 
 logger = logging.getLogger(__name__)
@@ -42,51 +40,3 @@ class TempFileList(list):
 
     def __del__(self):
         self.cleanup()
-
-
-class RunProcess(threading.Thread):
-    """
-    Run process and terminate a process given a specified time interval.
-
-    Adapted from: http://stackoverflow/questions/4158502/
-
-    Two input arguments:
-        - cmd, the command to execute in subprocess, as a list
-        - timeout, the maximum length of time to execute, in seconds
-
-    Example call:
-        - RunProcess(["./foo", "arg1"]).call()
-    """
-
-    def __init__(self, cmd, timeout=60):
-        super().__init__()
-        self.cmd = cmd
-        self.timeout = timeout
-
-    def run(self):
-        self.p = Popen(self.cmd)
-        self.p.wait()
-
-    def handle_failure(self):
-        pass
-
-    def call(self):
-        self.start()
-        self.join(self.timeout)
-        if self.is_alive():
-            self.p.terminate()
-            self.handle_failure()
-            self.join()
-
-
-class RunBMDS(RunProcess):
-    """
-    Custom thread class for running BMDS models. Logs failures when BMDS
-    models are killed, including printing the (d) file.
-    """
-
-    def handle_failure(self):
-        exe = self.cmd[0]
-        with open(self.cmd[1], 'r') as f:
-            dfile = f.read()
-        logger.warning('BMDS model killed: {}\n{}'.format(exe, dfile))